requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
schedule==1.2.1
python-dateutil==2.8.2
croniter==2.0.3
requests==2.32.3
